Prevent crash when event could not be encoded
Adds try/rescue around the json_encoder.encode_to_iodata!/1 function
call, which crashes when event contains data for which the json_encoder
(in our case Jason) has no explicit protocol defined.

This construct was added to be able to inspect the event that causes the
crash. Ideally we remove the try/rescue once we've been able to identify
and fix the root cause.